[MOD] cart_update_json

+++ ecommerce_product/controllers/website_sale.py
@@ -70,7 +70,7 @@ class WebsiteSaleProducts(ProductsFilter):
         parent_pack = request.env['product.product'].search([('name', '=', 'Open Pack')], limit=1)
         # parent_pack = request.env['product.product'].search([('id', '=', 1262)], limit=1)
         # if parent_pack and product_id == parent_pack.id and (set_qty == 0 or (add_qty and values['quantity'] == 0)):
-        if parent_pack and (set_qty == 0 or (add_qty and values['quantity'] == 0)):
+        if parent_pack and product_id == parent_pack.id and (set_qty == 0 or (add_qty and values['quantity'] == 0)):
             swatches_lines = order.order_line.filtered(lambda l: l.product_id.producttype_id.name == "Swatches")
             for line in swatches_lines:
                 line.unlink()
